working pairs allocation
Can now process pairs allocation.

The code analyses the preferences, and generates all possible pairs of
students who share at least one preference in their top three. Then it
creates all the sets of pairings, and for each set of pairings, we
analyse the cost matrix. At the end we select the lowest cost (highest
satisfaction).

The pairs is implemented entirely separately from the solo, which is
silly (thanks claude). But we should be able to refactor so that the
reading of arguments, the parsing of the csv and the output to csv/html
is unified.